Fix bug preventing authentication in certain cases
The src/io/utils/json_parse.js file is a pure Javascript JSON parser,
which allows us to parse the Terasology configuration files which
contain Java BigIntegers serialized as JSON integers, which are not
supported by the standard JSON.parse() function provided by browsers.

When I initially adapted the original json_parse.js by Douglas
Crockford (https://github.com/douglascrockford/JSON-js) by adding
support for jsbn BigIntegers, I forgot to consider the case of negative
numbers, and nothing ever failed until today when I tried to
authenticate and one of the BigIntegers in the keys generated by
Terasology happened to be negative, and the web frontend failed to
properly parse the config file pasted in the textbox (the regex wasn't
matching the number, thus the parser was falling back to interpret it as
a regular JS number, which is not able to hold such large numbers,
instead of building the proper BigInteger object).

So, I fixed the regex to match negative numbers as well and now it works
as expected.

Note that this bug was only affecting logins made by pasting the
configuration file in the web UI, because in communications with the
indentity storage server the BigIntegers are encoded as base64 strings
to save some bandwidth.